fairly confident final (fallback here if tweaking)
I've done a great deal of styling here to achieve the general idea, but matching the photographs perfectly is a challenge without more information about specifics that I presume would be provided by a designer. It was fun to learn about all the possibilities and having the freedom to size things for different resolutions.